Redesign from the ground up
In the spirit of 6.1 returning the Unreal trials, I wanted to pretty up this site. There were numerous issues with the previous version and I decided to scrap the whole thing. This version should hopefully be significantly better now that I've taken a course in web design. Not perfect but at least better.